refactor(executors): extract pure JSONL stream translation from huggingchat (812->594)
Port upstream #6016. Pure code motion into open-sse/executors/huggingchat/jsonlStream.ts leaf; huggingchat.ts re-imports. Verified no fork-only logic reverted. Regression guard: tests/unit/huggingchat-jsonl-split.test.ts.
